Build with Make

+++ common/os/linux/memory.hpp
@@ -1,12 +1,14 @@
 #pragma once
 
+#include <defines.hpp>
+
 namespace memory {
 namespace internal {
 
 
-void *allocate_pages(u64 size);
-void *allocate_pages(void *base_address, u64 size);
-void  free_pages(void *memory, u64 size);
+void *allocate_pages(usize size);
+void *allocate_pages(void *base_address, usize size);
+void  free_pages(void *memory, usize size);
 
 
 } // namespace internal
